Update setup_db fixture
setup_db now uses DatabaseConnector to create tables, so new
changes to table creation can be done only in one place. Also
the population querys in setup_db are more concise.

test_insert_new_book_existing_folder also had minor changes. Mainly
the name of the test method which was test_insert_book. It now
checks if the folder was NOT added, as it is a duplicate.