Add load_folder method to FolderDBHandler
This method may receives one sorting_key or/and a filter_key. The
valid values for sorting_key are name, added_date and path, and for
the filter_key are active and not_active, regarding the folder
presence on the file system (in the case of not pluged external
hds).